+++ src/mqtt_client/session/handshake.rs
@@ -64,7 +64,7 @@ where
                 connection,
                 &Connect {
                     keepalive,
-                    properties: Properties::Slice(&properties),
+                    properties: Properties::from_slice(&properties),
                     client_id: Utf8String(client_id.as_str()),
                     auth,
                     will,
@@ -128,7 +128,7 @@ where
         let mut keepalive_interval = self.runtime.keepalive_interval;
         let mut assigned_client_id = None;
 
-        for property in ack.properties.into_iter() {
+        for property in ack.properties.iter_concrete() {
             match match property {
                 Ok(property) => property,
                 Err(err) => {
@@ -138,7 +138,7 @@ where
             } {
                 Property::MaximumPacketSize(size) => maximum_packet_size = Some(size),
                 Property::AssignedClientIdentifier(id) => {
-                    assigned_client_id = Some(match id.0.try_into() {
+                    assigned_client_id = Some(match id.try_into() {
                         Ok(client_id) => client_id,
                         Err(_) => {
                             self.handle_disconnect();